Refactor to put pages under src and rename index
