+++ src/pages/employment_prediction.py
@@ -7,6 +7,7 @@ from dash import dcc
 from dash import html
 import dash_bootstrap_components as dbc
 from dash.dependencies import Input, Output, State
+import dash_loading_spinners
 import plotly.graph_objs as go
 import plotly.io as pio
 import plotly.express as px
@@ -19,10 +20,9 @@ if src_path not in sys.path:
     sys.path.append(src_path)
 
 from app import app
-from pages.layouts import header, footer
+from src.pages.layouts import header, footer, blank_placeholder_plot
 from src.data.gcp_strava_data_load_preprocess import load_strava_activity_data_from_bq
 from src.data.strava_data_load_preprocess import (
-    load_strava_activity_data,
     load_employment_model_data,
     preprocess_employment_model_data,
 )
@@ -30,7 +30,6 @@ from src.visualizations.employment_dash_plots import (
     plot_lgbm_model_predictions,
     plot_logreg_model_predictions,
     plot_weekly_start_times,
-    plot_eda_data,
     plot_training_data,
 )
 
@@ -47,183 +46,13 @@ except ImportError:
 # What's the definition of "work hours", activities are grouped down 9-11 is any activities started 9:00 - 11:59
 work_hours = [[9, 11], [13, 16]]
 
+loading_speed_multiplier = 1.5
+loading_color = "#e95420"
+loading_width = 125
+
+
 # ------------ END CONSTANTS -----------------------------#
 
-# for real time running on GCP App engine pull data from BQ
-raw_files_dict = load_strava_activity_data_from_bq()
-
-# # Use following code to test with sample data locally
-# # data_file_path = os.path.abspath(os.path.join(os.getcwd(), "data"))
-# # print("Loading Strava Data: " + data_file_path)
-# # # raw_files_dict = load_strava_activity_data(data_file_path)
-
-num_activities = len(raw_files_dict["TyAndrews"].type)
-
-activity_over_time = plot_eda_data(
-    raw_files_dict["TyAndrews"], [2013, 2023], "distance_raw_km", "Month"
-)
-
-# # page intro card for top of page under header and link to blog post on it.
-# jumbotron = dbc.Container(
-#     [
-#         html.H6(
-#             "Millions of people upload their activities to Strava every day. This got me wondering...",
-#             className="display-6",
-#         ),
-#         html.Hr(className="my-2"),
-#         html.H5(
-#             "What does Strava know about us from our uploaded activities?",
-#             className="display-5",
-#         ),
-#         html.P(
-#             "For a full description of the analysis below see the full write up here."
-#         ),
-#         html.P(
-#             dbc.Button(
-#                 "LEARN MORE",
-#                 size="lg",
-#                 href="https://ty-andrews.com/post/2021-02-23-what-does-strava-know-about-me/",
-#                 color="primary",
-#                 target="_blank",
-#             )
-#         ),  # , className="lead"
-#     ],
-#     className="h-100 p-5 bg-light border rounded-3",  # "dash-bootstrap py-3",
-#     fluid=True,
-# )
-
-activity_controls = dbc.Col(
-    [
-        html.P(
-            "All of my past 8 years of activities uploaded to Strava were exported. "
-            "From 2014-2017 I was training and going to school. 2017 to now I have been working "
-            "full time."
-        ),
-        dbc.Row(
-            [
-                dbc.Label("Y variable"),
-                dcc.Dropdown(
-                    id="eda-variable",
-                    options=[
-                        # {"label": col, "value": col} for col in raw_files_dict['TyAndrews'].columns
-                        {"label": col[0], "value": col[1]}
-                        for col in [
-                            ["Activity Distance", "distance_raw_km"],
-                            ["Elapsed Time", "elapsed_time_raw_hrs"],
-                            ["Moving Time", "moving_time_raw_hrs"],
-                            ["Elevation Gain", "elevation_gain"],
-                        ]
-                    ],
-                    value="distance_raw_km",
-                ),
-            ]
-        ),
-        dbc.Row(
-            [
-                dbc.Label("Display Per:"),
-                dcc.Dropdown(
-                    id="eda-group",
-                    options=[
-                        # {"label": col, "value": col} for col in raw_files_dict['TyAndrews'].columns
-                        {"label": col, "value": col}
-                        for col in ["Day", "Week", "Month", "Quarter", "Year"]
-                    ],
-                    value="Month",
-                ),
-            ]
-        ),
-        dbc.Row(
-            [
-                dbc.Label("Year Range"),
-                dcc.RangeSlider(
-                    id="year-selector",
-                    min=2013,
-                    max=2023,
-                    step=1,
-                    value=[2013, 2023],
-                    marks={
-                        2013: "2013",
-                        2014: "'14",
-                        2015: "'15",
-                        2016: "'16",
-                        2017: "'17",
-                        2018: "'18",
-                        2019: "'19",
-                        2020: "'20",
-                        2021: "'21",
-                        2022: "'22",
-                        2023: "2023",
-                    },
-                ),
-            ]
-        ),
-    ],
-    # body=True,
-)
-
-# DCC card for showing my activities over time etc.
-data_intro_card = dbc.Card(
-    [
-        dbc.CardHeader(html.H4("Explore My Strava Data")),
-        dbc.CardBody(
-            dbc.Row(
-                [
-                    dbc.Col(activity_controls, width=12, lg=3),
-                    dbc.Col(
-                        dbc.Card(
-                            dbc.CardBody(
-                                dcc.Loading(
-                                    children=dcc.Graph(
-                                        id="eda-plot", figure=activity_over_time
-                                    ),
-                                    type="cube",
-                                    color="#e95420",
-                                )
-                            )
-                        ),
-                        width=12,
-                        lg=9,
-                    ),
-                ]
-            )
-        ),
-    ],
-    color="#E0E0E0",
-)
-
-# pre-generate all plots for building th core components
-unemployed_activity_start_time_fig = plot_weekly_start_times(
-    raw_files_dict["TyAndrews"], 2014, 2017, work_hours, "Training + Studying"
-)
-employed_activity_start_time_fig = plot_weekly_start_times(
-    raw_files_dict["TyAndrews"], 2018, 2020, work_hours, "Working Full Time"
-)
-
-train_employment_data, test_employment_data = load_employment_model_data()
-
-train_data, train_labels = preprocess_employment_model_data(
-    train_employment_data, work_hours
-)
-test_data, test_labels = preprocess_employment_model_data(
-    test_employment_data, work_hours
-)
-
-training_data_plot = plot_training_data(
-    train_data, train_labels, test_data, test_labels
-)
-
-lgbm_train_plot = plot_lgbm_model_predictions(
-    train_data[["morn", "aft"]], train_data[["morn", "aft"]], train_labels, "train"
-)  # optional: lgbm_model,
-lgbm_test_plot = plot_lgbm_model_predictions(
-    test_data[["morn", "aft"]], test_data[["morn", "aft"]], test_labels, "test"
-)
-logreg_train_plot = plot_logreg_model_predictions(
-    train_data[["morn", "aft"]], train_labels, "train"
-)  # Optional: logreg_model,
-logreg_test_plot = plot_logreg_model_predictions(
-    test_data[["morn", "aft"]], test_labels, "test"
-)
 
 # card holding the employment status hypothesis details and results plots.
 employment_hypoth = dbc.Card(
@@ -264,23 +93,33 @@ employment_hypoth = dbc.Card(
                                         dbc.Row(
                                             [
                                                 dbc.Col(
-                                                    dcc.Graph(
-                                                        id="unempl-explore-plot",
-                                                        figure=unemployed_activity_start_time_fig,
-                                                        config={
-                                                            "displayModeBar": False
-                                                        },
+                                                    dash_loading_spinners.Ring(
+                                                        children=dcc.Graph(
+                                                            id="unempl-explore-plot",
+                                                            figure=blank_placeholder_plot(),
+                                                            config={
+                                                                "displayModeBar": False
+                                                            },
+                                                        ),
+                                                        speed_multiplier=loading_speed_multiplier,
+                                                        width=loading_width,
+                                                        color=loading_color,
                                                     ),
                                                     width=12,
                                                     lg=6,
                                                 ),
                                                 dbc.Col(
-                                                    dcc.Graph(
-                                                        id="empl-explore-plot",
-                                                        figure=employed_activity_start_time_fig,
-                                                        config={
-                                                            "displayModeBar": False
-                                                        },
+                                                    dash_loading_spinners.Ring(
+                                                        children=dcc.Graph(
+                                                            id="empl-explore-plot",
+                                                            figure=blank_placeholder_plot(),
+                                                            config={
+                                                                "displayModeBar": False
+                                                            },
+                                                        ),
+                                                        speed_multiplier=loading_speed_multiplier,
+                                                        width=loading_width,
+                                                        color=loading_color,
                                                     ),
                                                     width=12,
                                                     lg=6,
@@ -326,16 +165,20 @@ employment_hypoth = dbc.Card(
                                         dbc.Row(
                                             [
                                                 dbc.Col(
-                                                    dcc.Graph(
-                                                        id="train-data-plot",
-                                                        figure=training_data_plot,
-                                                        config={
-                                                            "displayModeBar": False
-                                                        },
+                                                    dash_loading_spinners.Ring(
+                                                        dcc.Graph(
+                                                            id="train-data-plot",
+                                                            figure=blank_placeholder_plot(),
+                                                            config={
+                                                                "displayModeBar": False
+                                                            },
+                                                        ),
+                                                        speed_multiplier=loading_speed_multiplier,
+                                                        width=loading_width,
+                                                        color=loading_color,
                                                     ),
                                                     width=12,
                                                 ),
-                                                # dbc.Col(dcc.Graph(id='empl-explore-plot', figure=employed_activity_start_time_fig), width = 6)])
                                             ]
                                         )
                                     ]
@@ -390,13 +233,20 @@ employment_hypoth = dbc.Card(
                                                                 "Training Data - 70%"
                                                             ),
                                                         ),
-                                                        dcc.Graph(
-                                                            id="logreg-train-data-plot",
-                                                            figure=logreg_train_plot,
-                                                            style={"height": "350px"},
-                                                            config={
-                                                                "displayModeBar": False
-                                                            },
+                                                        dash_loading_spinners.Ring(
+                                                            dcc.Graph(
+                                                                id="logreg-train-data-plot",
+                                                                figure=blank_placeholder_plot(),
+                                                                style={
+                                                                    "height": "350px"
+                                                                },
+                                                                config={
+                                                                    "displayModeBar": False
+                                                                },
+                                                            ),
+                                                            speed_multiplier=loading_speed_multiplier,
+                                                            width=loading_width,
+                                                            color=loading_color,
                                                         ),
                                                     ],
                                                     width=12,
@@ -407,13 +257,20 @@ employment_hypoth = dbc.Card(
                                                         html.Center(
                                                             html.H5("Test Data - 30%"),
                                                         ),
-                                                        dcc.Graph(
-                                                            id="logreg-test-data-plot",
-                                                            figure=logreg_test_plot,
-                                                            style={"height": "350px"},
-                                                            config={
-                                                                "displayModeBar": False
-                                                            },
+                                                        dash_loading_spinners.Ring(
+                                                            children=dcc.Graph(
+                                                                id="logreg-test-data-plot",
+                                                                figure=blank_placeholder_plot(),
+                                                                style={
+                                                                    "height": "350px"
+                                                                },
+                                                                config={
+                                                                    "displayModeBar": False
+                                                                },
+                                                            ),
+                                                            speed_multiplier=loading_speed_multiplier,
+                                                            width=loading_width,
+                                                            color=loading_color,
                                                         ),
                                                     ],
                                                     width=12,
@@ -426,25 +283,35 @@ employment_hypoth = dbc.Card(
                                         dbc.Row(
                                             [
                                                 dbc.Col(
-                                                    dcc.Graph(
-                                                        id="lgbm-train-data-plot",
-                                                        figure=lgbm_train_plot,
-                                                        style={"height": "350px"},
-                                                        config={
-                                                            "displayModeBar": False
-                                                        },
+                                                    dash_loading_spinners.Ring(
+                                                        children=dcc.Graph(
+                                                            id="lgbm-train-data-plot",
+                                                            figure=blank_placeholder_plot(),
+                                                            style={"height": "350px"},
+                                                            config={
+                                                                "displayModeBar": False
+                                                            },
+                                                        ),
+                                                        speed_multiplier=loading_speed_multiplier,
+                                                        width=loading_width,
+                                                        color=loading_color,
                                                     ),
                                                     width=12,
                                                     lg=6,
                                                 ),
                                                 dbc.Col(
-                                                    dcc.Graph(
-                                                        id="lgbm-test-data-plot",
-                                                        figure=lgbm_test_plot,
-                                                        style={"height": "350px"},
-                                                        config={
-                                                            "displayModeBar": False
-                                                        },
+                                                    dash_loading_spinners.Ring(
+                                                        children=dcc.Graph(
+                                                            id="lgbm-test-data-plot",
+                                                            figure=blank_placeholder_plot(),
+                                                            style={"height": "350px"},
+                                                            config={
+                                                                "displayModeBar": False
+                                                            },
+                                                        ),
+                                                        speed_multiplier=loading_speed_multiplier,
+                                                        width=loading_width,
+                                                        color=loading_color,
                                                     ),
                                                     width=12,
                                                     lg=6,
@@ -477,17 +344,96 @@ layout = html.Div(
             )
         ),
         footer(),
+        html.Div(id="employ-page-load-div"),
     ],
 )
 
-# Updating the eda plot based on slider/dropdown selections
-# @app.callback(
-#     Output("eda-plot", "figure"),
-#     [
-#         Input("eda-variable", "value"),
-#         Input("year-selector", "value"),
-#         Input("eda-group", "value"),
-#     ],
-# )
-# def update_eda_plot(y_label, year_range, group_by):
-#     return plot_eda_data(raw_files_dict["TyAndrews"], year_range, y_label, group_by)
+# On page load, populate first necessary data/plots
+@app.callback(
+    Output("unempl-explore-plot", "figure"),
+    Output("empl-explore-plot", "figure"),
+    Output("train-data-plot", "figure"),
+    Output("logreg-train-data-plot", "figure"),
+    Output("logreg-test-data-plot", "figure"),
+    Output("lgbm-test-data-plot", "figure"),
+    Output("lgbm-train-data-plot", "figure"),
+    Input("employ-page-load-div", "children"),
+    [
+        State("strava-data", "data"),
+        State("empl-data", "data"),
+    ],
+    prevent_initial_call=False,
+)
+def load_app_data(_page_load, strava_dict, empl_dict):
+
+    if strava_dict is None:
+        strava_df = load_strava_activity_data_from_bq()["TyAndrews"]
+        strava_dict = strava_df.to_dict("records")
+    else:
+        strava_df = pd.DataFrame(strava_dict)
+
+    if empl_dict is None:
+        train_employment_data, test_employment_data = load_employment_model_data()
+
+        train_data, train_labels = preprocess_employment_model_data(
+            train_employment_data, work_hours
+        )
+        test_data, test_labels = preprocess_employment_model_data(
+            test_employment_data, work_hours
+        )
+
+        empl_dict = {
+            "train_data": train_data.to_dict(),
+            "train_labels": train_labels.to_dict(),
+            "test_data": test_data.to_dict(),
+            "test_labels": test_labels.to_dict(),
+        }
+    else:
+        train_data = pd.DataFrame(empl_dict["train_data"])
+        train_labels = pd.Series(empl_dict["train_labels"])
+        test_data = pd.DataFrame(empl_dict["test_data"])
+        test_labels = pd.Series(empl_dict["test_labels"])
+
+    train_data = pd.DataFrame(empl_dict["train_data"])
+    train_labels = pd.Series(empl_dict["train_labels"])
+    test_data = pd.DataFrame(empl_dict["test_data"])
+    test_labels = pd.Series(empl_dict["test_labels"])
+
+    training_data_plot = plot_training_data(
+        train_data, train_labels, test_data, test_labels
+    )
+
+    # pre-generate all plots for building th core components
+    unemployed_activity_start_time_fig = plot_weekly_start_times(
+        strava_df, 2014, 2017, work_hours, "Training + Studying"
+    )
+    employed_activity_start_time_fig = plot_weekly_start_times(
+        strava_df, 2018, 2020, work_hours, "Working Full Time"
+    )
+
+    training_data_plot = plot_training_data(
+        train_data, train_labels, test_data, test_labels
+    )
+
+    lgbm_train_plot = plot_lgbm_model_predictions(
+        train_data[["morn", "aft"]], train_data[["morn", "aft"]], train_labels, "train"
+    )  # optional: lgbm_model,
+    lgbm_test_plot = plot_lgbm_model_predictions(
+        test_data[["morn", "aft"]], test_data[["morn", "aft"]], test_labels, "test"
+    )
+    logreg_train_plot = plot_logreg_model_predictions(
+        train_data[["morn", "aft"]], train_labels, "train"
+    )  # Optional: logreg_model,
+    logreg_test_plot = plot_logreg_model_predictions(
+        test_data[["morn", "aft"]], test_labels, "test"
+    )
+
+    return (
+        unemployed_activity_start_time_fig,
+        employed_activity_start_time_fig,
+        training_data_plot,
+        logreg_train_plot,
+        logreg_test_plot,
+        lgbm_test_plot,
+        lgbm_train_plot,
+    )
