+++ src/pages/kudos_prediction.py
@@ -1,7 +1,9 @@
 import dash_bootstrap_components as dbc
 from dash import dcc
 from dash import html
-from dash.dependencies import Input, Output
+from dash.dependencies import Input, Output, State
+import dash_loading_spinners
+import pandas as pd
 import time
 import sys
 import os
@@ -11,7 +13,7 @@ if src_path not in sys.path:
     sys.path.append(src_path)
 
 from app import app
-from pages.layouts import header, footer
+from src.pages.layouts import header, footer, blank_placeholder_plot
 from src.visualizations.kudos_prediction_dash_plots import (
     plot_predicted_kudos_value,
     plot_distance_prediction,
@@ -30,53 +32,12 @@ custom_name_bool = 0
 achievements = 15
 elevation = 4  # 100's of meters
 
+loading_speed_multiplier = 1.5
+loading_color = "#e95420"
+loading_width = 125
+
 # date of raw data file to load
-static_kudos_predictions = load_static_kudos_predictions("2022-05-18")
-
-# for real time running on GCP App engine pull data from BQ
-# raw_files_dict = load_strava_activity_data_from_bq()
-
-
-kudos_prediction = (
-    static_kudos_predictions.loc[
-        (static_kudos_predictions.distance == distance)
-        & (static_kudos_predictions.achievements == achievements)
-        & (static_kudos_predictions.elevation == elevation * 100)
-        & (static_kudos_predictions.custom_name_bool == custom_name_bool),
-        "kudos_prediction",
-    ].iloc[0]
-    * num_followers
-)
-
-
-initial_kudos_plot = plot_predicted_kudos_value(round(kudos_prediction))
-distance_plot = plot_distance_prediction(
-    static_kudos_predictions,
-    kudos_prediction,
-    num_followers,
-    custom_name_bool,
-    distance,
-    elevation * 100,
-    achievements,
-)
-elevation_plot = plot_elevation_prediction(
-    static_kudos_predictions,
-    kudos_prediction,
-    num_followers,
-    custom_name_bool,
-    distance,
-    elevation * 100,
-    achievements,
-)
-achievement_plot = plot_achievement_prediction(
-    static_kudos_predictions,
-    kudos_prediction,
-    num_followers,
-    custom_name_bool,
-    distance,
-    elevation * 100,
-    achievements,
-)
+raw_kudos_data_date = "2022-05-18"
 
 kudos_controls = html.Div(
     [
@@ -150,16 +111,17 @@ kudos_prediction_card = dbc.Card(
                             dbc.Card(
                                 dbc.CardBody(
                                     children=[
-                                        dcc.Loading(
+                                        dash_loading_spinners.Ring(
                                             children=[
                                                 dcc.Graph(
                                                     id="kudos-plot",
-                                                    figure=initial_kudos_plot,
+                                                    figure=blank_placeholder_plot(),
                                                     config={"displayModeBar": False},
                                                 )
                                             ],
-                                            type="circle",
-                                            color="#e95420",
+                                            speed_multiplier=loading_speed_multiplier,
+                                            width=loading_width,
+                                            color=loading_color,
                                         ),
                                     ]
                                 )
@@ -184,18 +146,19 @@ kudos_prediction_card = dbc.Card(
                                     dbc.Card(
                                         dbc.CardBody(
                                             children=[
-                                                dcc.Loading(
+                                                dash_loading_spinners.Ring(
                                                     children=[
                                                         dcc.Graph(
                                                             id="dist-plot",
-                                                            figure=distance_plot,
+                                                            figure=blank_placeholder_plot(),
                                                             config={
                                                                 "displayModeBar": False
                                                             },
                                                         )
                                                     ],
-                                                    type="circle",
-                                                    color="#e95420",
+                                                    speed_multiplier=loading_speed_multiplier,
+                                                    width=loading_width,
+                                                    color=loading_color,
                                                 ),
                                             ]
                                         ),
@@ -203,18 +166,19 @@ kudos_prediction_card = dbc.Card(
                                     dbc.Card(
                                         dbc.CardBody(
                                             children=[
-                                                dcc.Loading(
+                                                dash_loading_spinners.Ring(
                                                     children=[
                                                         dcc.Graph(
                                                             id="elev-plot",
-                                                            figure=elevation_plot,
+                                                            figure=blank_placeholder_plot(),
                                                             config={
                                                                 "displayModeBar": False
                                                             },
                                                         )
                                                     ],
-                                                    type="circle",
-                                                    color="#e95420",
+                                                    speed_multiplier=loading_speed_multiplier,
+                                                    width=loading_width,
+                                                    color=loading_color,
                                                 ),
                                             ]
                                         )
@@ -222,18 +186,19 @@ kudos_prediction_card = dbc.Card(
                                     dbc.Card(
                                         dbc.CardBody(
                                             children=[
-                                                dcc.Loading(
+                                                dash_loading_spinners.Ring(
                                                     children=[
                                                         dcc.Graph(
                                                             id="achiev-plot",
-                                                            figure=achievement_plot,
+                                                            figure=blank_placeholder_plot(),
                                                             config={
                                                                 "displayModeBar": False
                                                             },
                                                         )
                                                     ],
-                                                    type="circle",
-                                                    color="#e95420",
+                                                    speed_multiplier=loading_speed_multiplier,
+                                                    width=loading_width,
+                                                    color=loading_color,
                                                 ),
                                             ]
                                         )
@@ -261,6 +226,8 @@ layout = html.Div(
             )
         ),
         footer(),
+        html.Div(id="kudos-page-load-div"),
+        dcc.Store(id="kudos-static-data", data=None, storage_type="session"),
     ]
 )
 
@@ -270,17 +237,33 @@ layout = html.Div(
     Output(component_id="dist-plot", component_property="figure"),
     Output(component_id="elev-plot", component_property="figure"),
     Output(component_id="achiev-plot", component_property="figure"),
+    Output("kudos-static-data", "data"),
     [
         Input(component_id="num-followers", component_property="value"),
         Input(component_id="custom-name-input", component_property="value"),
         Input(component_id="distance-slider", component_property="value"),
         Input(component_id="elevation-slider", component_property="value"),
         Input(component_id="achievements-slider", component_property="value"),
+        Input("kudos-page-load-div", "children"),
     ],
+    State("kudos-static-data", "kudos_static_dict"),
 )
 def update_predicted_kudos_number(
-    num_followers, custom_name_bool, distance, elevation, achievements
+    num_followers,
+    custom_name_bool,
+    distance,
+    elevation,
+    achievements,
+    _children,
+    kudos_static_dict,
 ):
+
+    if kudos_static_dict is None:
+        static_kudos_predictions = load_static_kudos_predictions(raw_kudos_data_date)
+        kudos_static_dict = static_kudos_predictions.to_dict("records")
+    else:
+        static_kudos_predictions = pd.DataFrame(kudos_static_dict)
+
     elevation_m = elevation * 100  # get into meters
 
     perc_followers = static_kudos_predictions.loc[
@@ -329,4 +312,5 @@ def update_predicted_kudos_number(
         distance_value_fig,
         elevation_value_fig,
         achievement_value_fig,
+        kudos_static_dict,
     )
